[AV2-228] created new component for shared contact form. moved methods and html, css things to new component.
