+++ src/app/shared/components/contact-number-form/contact-number-form.component.html
@@ -1,0 +1,36 @@
+<ng-container *ngIf="page === 'settings'">
+  <ion-item>
+    <ion-label>Select Country</ion-label>
+    <ion-select [(ngModel)]="countryModel" (ionChange)="updateCountry()">
+      <ion-select-option *ngFor="let countryCode of contact.countryCodes" [value]="countryCode.code">{{countryCode.name}}</ion-select-option>
+    </ion-select>
+  </ion-item>
+  <ion-item>
+    <ion-grid class="contact-grid">
+      <ion-row>
+        <ion-col class="contact-label">
+            <ion-label>Contact No.</ion-label>
+        </ion-col>
+      </ion-row>
+      <ion-row class="text-field" nowrap align-items-center>
+        <ion-col>
+          <ion-row class="contact-details">
+            <ion-col size="4">
+              <p class="country-code">{{selectedCountryCode}}</p>
+            </ion-col>
+            <ion-col>
+                <input maxlength="12" pattern="{{activeContactPattern}}" placeholder="{{activeContactPlaceholder}}" [(ngModel)]="contactNumber" type="text"
+                class="contact-input" (keydown)="disableArrowKeys($event)" (keyup)="formatContactNumber()" />
+            </ion-col>
+          </ion-row>
+        </ion-col>
+        <ion-col size="5">
+          <ion-button class="update-btn" [disabled]="updating" (click)="updateContactNumber()">
+            <ng-container *ngIf="!updating">Update</ng-container>
+            <ng-container *ngIf="updating">Updating..</ng-container>
+          </ion-button>
+        </ion-col>
+      </ion-row>
+    </ion-grid>
+  </ion-item>
+</ng-container>
