+++ src/app/shared/components/contact-number-form/contact-number-form.component.scss
@@ -1,0 +1,56 @@
+.btn-right {
+    position: absolute;
+    right: 10px;
+  }
+  
+  .contact-grid {
+    margin: 0px;
+    width: 100%;
+    padding: 0px;
+  
+    .contact-label {
+      padding-bottom: 0px;
+      padding-left: 0;
+      padding-right: 0;
+    }
+  
+    ion-row {
+      &.text-field {
+        ion-col {
+          padding-top: 0px;
+          padding-left: 0;
+          padding-right: 0;
+          text-align: center;
+  
+          p {
+            &.country-code {
+              margin-bottom: 0;
+              margin-top: 10px;
+              font-weight: bold;
+            }
+          }
+  
+          &.contact-details {
+            padding-bottom: 0;
+            padding-left: 0;
+          }
+        }
+      }
+    }
+  }
+  
+  .contact-input {
+    border-top: 0;
+    border-left: 0;
+    border-right: 0;
+    border-bottom: 1px;
+    border-style: solid;
+    border-color: lightgrey;
+    margin-top: 8px;
+    width: 100%;
+    color: var(--practera-label-color);
+    &:focus {
+      outline: none;
+    }
+  }
+
