+++ src/app/shared/components/contact-number-form/contact-number-form.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ContactNumberFormComponent } from './contact-number-form.component';
+
+describe('ContactNumberFormComponent', () => {
+  let component: ContactNumberFormComponent;
+  let fixture: ComponentFixture<ContactNumberFormComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ContactNumberFormComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ContactNumberFormComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
